+++ shared/framework-kernel/src/declarations/framework/kernel.ts
@@ -8,6 +8,7 @@ export interface KernelOptions {
   config: HttpConfig;
   securityKey: string;
   redirectUri: string;
+  tenantId: string;
   isUseCrypto: boolean;
   isAutoRefreshToken: boolean;
 }