+++ shared/bpmn-designer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @herodotus-cloud/bpmn-designer
 
+## 1.7.6
+
+### Patch Changes
+
+- 迁移 Pinia Helper 类
+- Updated dependencies
+  - @herodotus-cloud/components@1.7.6
+  - @herodotus-cloud/bpmn-apis@1.7.6
+  - @herodotus-cloud/form-apis@1.7.6
+  - @herodotus-cloud/core@1.7.6
+
 ## 1.7.5
 
 ### Patch Changes