+++ shared/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @herodotus-cloud/core
 
+## 1.7.6
+
+### Patch Changes
+
+- 迁移 Pinia Helper 类
+
 ## 1.7.5
 
 ### Patch Changes